remove archived modules from the project
This is mainly due to this modules are generating errors by building
on other platforms rather than UNIX or POSIX based platforms, like,
**cof, cof** Windows **cof** . This modules need big improvements and
probably a full rewrite which will cause some breaking changes.

So, it's preferred stable and maintained modules that can be built
on the major supported platforms (Linux, Windows, macOS).

This modules will still be available on 'archived-modules' branch
within an undefined period of time.

Signed-off-by: Luís Ferreira <contact@lsferreira.net>